// src/models/debilidades.model.ts
export interface Debilidades {
    tipo: string; // Forgeign Key UUID
    debilidad_1: string; // Foreign Key UUID referencing tipos
    debilidad_2: string; // Foreign Key UUID referencing tipos
    debilidad_3: string; // Foreign Key UUID referencing tipos
    debilidad_4: string; // Foreign Key UUID referencing tipos
    debilidad_5: string; // Foreign Key UUID referencing tipos
    key_debilidades: string; // Primary Key UUID
}